pyproject.toml
[tool.poetry]
name = "tgmusicbot"
version = "0.1.0"
description = ""
authors = ["Actus Nileh <actusnileh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.7.0"
yt-dlp = "^2024.5.27"
youtube-search = "^2.1.2"
django-environ = "^0.11.2"
pydantic-settings = "^2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
